Fix aspect ratio
This takes advantage of the fact that ImageMagick will keep scales
aspect-ratio-correct by default. So I just pass it the size of the
label for geometry and it finds the best fit for the image.

There's also a bit of work on simplifying the transfer of data
into Tcl using a LinkHandle, and although it's a bit easier on the
eyes now, it's still just as slow. I think I'll need to add -data
support to C++/Tk itself.